Reorganize source files and packages
