+++ .github/workflows/flutter-test.yml
@@ -1,0 +1,30 @@
+name: flutter build and test
+on:
+  push:
+    branches:
+      - master
+  pull_request:
+
+jobs:
+  test:
+    name: Test on ${{ matrix.os }}
+    runs-on: ${{ matrix.os }}
+    strategy:
+      matrix:
+        os: [ubuntu-latest]
+    env:
+      working-directory: ./app
+    steps:
+      - uses: actions/checkout@v1
+      - uses: actions/setup-java@v1
+        with:
+          java-version: "12.x"
+      - uses: subosito/flutter-action@v1
+        with:
+          flutter-version: "1.12.13+hotfix.8"
+      - run: flutter pub get
+        working-directory: ${{env.working-directory}}
+      - run: flutter test
+        working-directory: ${{env.working-directory}}
+      - run: flutter build apk
+        working-directory: ${{env.working-directory}}
