+++ app/lib/models/feed_comment.dart
@@ -1,0 +1,11 @@
+class FeedCommentModel {
+  final String author;
+  final String avatarUrl;
+  final String text;
+
+  FeedCommentModel({
+    this.author,
+    this.text,
+    this.avatarUrl,
+  });
+}
