+++ app/lib/models/profile.dart
@@ -1,0 +1,11 @@
+import 'package:eventapp/models/profile_post.dart';
+
+class ProfileModel {
+  final String name;
+  final List<ProfilePostModel> posts;
+
+  ProfileModel({
+    this.name,
+    this.posts,
+  });
+}
