+++ app/lib/models/profile_post.dart
@@ -1,0 +1,9 @@
+class ProfilePostModel {
+  final String imageUrl;
+  final String text;
+
+  ProfilePostModel({
+    this.imageUrl,
+    this.text
+  });
+}
