+++ app/lib/models/schedule_event.dart
@@ -1,0 +1,15 @@
+class ScheduleEvent {
+  final DateTime startTime;
+  final DateTime endTime;
+  final String pictureUrl;
+  final String title;
+  final String text;
+
+  ScheduleEvent({
+    this.startTime,
+    this.endTime,
+    this.pictureUrl,
+    this.title,
+    this.text,
+  });
+}
